Correct the way it looks for the head.
It was looking for the head without taking into account any previous
information about it. This meant that sometimes the head could be mistaken
by e.g. a hand close to the chest if that hand/extrema was evaluated first.
This adds the information about the head joint from the previous frame so
it evaluates the extrema that is closer to that which results in a correction.