Update skeltrack-skeleton.c
Made changes to handle the cases when hand gets close to torso and thus their nodes get connected and so previously hand was returned as null. Now in such cases when hand is NULL previous two frames are looked and using linear extrapolation the new position is calculated.
Also handled the case when hands were crossed ie. when right hand is on left side of left shoulder and vice versa.(Earlier it connected right hand to left shoulder in such case).